fixed int writable

+++ step2/topN/PredictionMapper.java
@@ -18,7 +18,7 @@ import org.apache.hadoop.mapreduce.Mapper.Context;
 
 import algo1.job1.Pair;
 
-public class PredictionMapper extends Mapper<IntWritable, Text, LongWritable, IDSIWritable> {
+public class PredictionMapper extends Mapper<LongWritable, Text, IntWritable, IDSIWritable> {
 
 	
 	// the position of the key in the line
@@ -31,7 +31,7 @@ public class PredictionMapper extends Mapper<IntWritable, Text, LongWritable, ID
 	
 	private static final String SIDE_DATA = "TestingRatings.txt";
 	
-	private final LongWritable keyOut = new LongWritable();
+	private final IntWritable keyOut = new IntWritable();
 	private final IDSIWritable valOut = new IDSIWritable();
 
 	private final Set<Integer> testUserID = new HashSet<Integer>();
@@ -68,7 +68,7 @@ public class PredictionMapper extends Mapper<IntWritable, Text, LongWritable, ID
 	 * emits either movieid or userid for the key, and the rating for the value
 	 */
 	@Override
-	public void map(IntWritable key, Text value, Context context) throws IOException, InterruptedException {
+	public void map(LongWritable key, Text value, Context context) throws IOException, InterruptedException {
 		String record = value.toString();
 		
 		String[] kv = record.split("\t");
